refactor: 회원 도메인 패키지 이동
