+++ src/main/java/shop/tryit/config/WebSecurityConfig.java
@@ -15,7 +15,7 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
-import shop.tryit.domain.member.MemberSecurityService;
+import shop.tryit.domain.member.service.MemberSecurityService;
 
 @EnableWebSecurity
 @RequiredArgsConstructor
