+++ src/main/java/shop/tryit/domain/cart/entity/Cart.java
@@ -13,7 +13,7 @@ import javax.persistence.OneToOne;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import shop.tryit.domain.common.BaseTimeEntity;
-import shop.tryit.domain.member.Member;
+import shop.tryit.domain.member.entity.Member;
 
 @Entity
 @Getter
