+++ src/main/java/shop/tryit/domain/cart/repository/CartRepository.java
@@ -1,7 +1,7 @@
 package shop.tryit.domain.cart.repository;
 
 import shop.tryit.domain.cart.entity.Cart;
-import shop.tryit.domain.member.Member;
+import shop.tryit.domain.member.entity.Member;
 
 public interface CartRepository {
