+++ src/main/java/shop/tryit/domain/member/service/MemberService.java
@@ -1,9 +1,11 @@
-package shop.tryit.domain.member;
+package shop.tryit.domain.member.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import shop.tryit.domain.member.entity.Member;
+import shop.tryit.domain.member.repository.MemberRepository;
 
 @Service
 @Transactional(readOnly = true)
