+++ src/main/java/shop/tryit/domain/notice/service/NoticeFacade.java
@@ -5,8 +5,8 @@ import org.springframework.data.domain.Page;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
-import shop.tryit.domain.member.Member;
-import shop.tryit.domain.member.MemberService;
+import shop.tryit.domain.member.entity.Member;
+import shop.tryit.domain.member.service.MemberService;
 import shop.tryit.domain.notice.dto.NoticeSaveFormDto;
 import shop.tryit.domain.notice.dto.NoticeSearchFormDto;
 import shop.tryit.domain.notice.dto.NoticeUpdateFormDto;
