+++ src/main/java/shop/tryit/domain/order/entity/Order.java
@@ -18,7 +18,7 @@ import javax.persistence.Table;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import shop.tryit.domain.common.BaseTimeEntity;
-import shop.tryit.domain.member.Member;
+import shop.tryit.domain.member.entity.Member;
 
 @Entity
 @Getter
