+++ src/main/java/shop/tryit/domain/order/repository/OrderRepository.java
@@ -3,7 +3,7 @@ package shop.tryit.domain.order.repository;
 import java.util.Optional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import shop.tryit.domain.member.Member;
+import shop.tryit.domain.member.entity.Member;
 import shop.tryit.domain.order.dto.OrderSearchDto;
 import shop.tryit.domain.order.entity.Order;
