+++ src/main/java/shop/tryit/domain/order/service/OrderFacade.java
@@ -13,8 +13,8 @@ import shop.tryit.domain.cart.service.CartItemService;
 import shop.tryit.domain.cart.service.CartService;
 import shop.tryit.domain.item.entity.Item;
 import shop.tryit.domain.item.service.ItemService;
-import shop.tryit.domain.member.Member;
-import shop.tryit.domain.member.MemberService;
+import shop.tryit.domain.member.entity.Member;
+import shop.tryit.domain.member.service.MemberService;
 import shop.tryit.domain.order.dto.OrderDetailDto;
 import shop.tryit.domain.order.dto.OrderFindDto;
 import shop.tryit.domain.order.dto.OrderFindDto.OrderDetailFindDto;
