+++ src/main/java/shop/tryit/domain/payment/PaymentFacade.java
@@ -12,8 +12,8 @@ import shop.tryit.domain.item.entity.Image;
 import shop.tryit.domain.item.entity.Item;
 import shop.tryit.domain.item.service.ImageService;
 import shop.tryit.domain.item.service.ItemService;
-import shop.tryit.domain.member.Member;
-import shop.tryit.domain.member.MemberService;
+import shop.tryit.domain.member.entity.Member;
+import shop.tryit.domain.member.service.MemberService;
 import shop.tryit.domain.payment.dto.PaymentRequestDto;
 import shop.tryit.domain.payment.dto.PaymentResponseDto;
 import shop.tryit.domain.payment.dto.PaymentSaveDto;
