+++ src/main/java/shop/tryit/domain/question/service/QnAFacade.java
@@ -10,8 +10,8 @@ import org.springframework.transaction.annotation.Transactional;
 import shop.tryit.domain.answer.dto.AnswerFormDto;
 import shop.tryit.domain.answer.entity.Answer;
 import shop.tryit.domain.answer.service.AnswerService;
-import shop.tryit.domain.member.Member;
-import shop.tryit.domain.member.MemberService;
+import shop.tryit.domain.member.entity.Member;
+import shop.tryit.domain.member.service.MemberService;
 import shop.tryit.domain.question.dto.QuestionCheckPasswordFormDto;
 import shop.tryit.domain.question.dto.QuestionFormDto;
 import shop.tryit.domain.question.dto.QuestionSaveFormDto;
