+++ src/main/java/shop/tryit/repository/member/MemberJpaRepository.java
@@ -2,11 +2,12 @@ package shop.tryit.repository.member;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
-import shop.tryit.domain.member.Member;
+import shop.tryit.domain.member.entity.Member;
 
 public interface MemberJpaRepository extends JpaRepository<Member, Long> {
 
     boolean existsByEmail(String email);
 
     Optional<Member> findByEmail(String email);
+
 }
