+++ src/main/java/shop/tryit/repository/order/OrderCustom.java
@@ -2,7 +2,7 @@ package shop.tryit.repository.order;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import shop.tryit.domain.member.Member;
+import shop.tryit.domain.member.entity.Member;
 import shop.tryit.domain.order.dto.OrderSearchDto;
 
 public interface OrderCustom {
