+++ src/main/java/shop/tryit/repository/order/OrderCustomImpl.java
@@ -1,6 +1,6 @@
 package shop.tryit.repository.order;
 
-import static shop.tryit.domain.member.QMember.member;
+import static shop.tryit.domain.member.entity.QMember.member;
 import static shop.tryit.domain.order.entity.QOrder.order;
 
 import com.querydsl.jpa.impl.JPAQuery;
@@ -10,7 +10,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.support.PageableExecutionUtils;
-import shop.tryit.domain.member.Member;
+import shop.tryit.domain.member.entity.Member;
 import shop.tryit.domain.order.dto.OrderDetailSearchDto;
 import shop.tryit.domain.order.dto.OrderSearchDto;
 import shop.tryit.domain.order.dto.QOrderSearchDto;
@@ -29,23 +29,9 @@ public class OrderCustomImpl implements OrderCustom {
     }
 
     private List<OrderSearchDto> orderSearchContents(Member findMember) {
-        QOrderSearchDto qOrderSearchDto =
-                new QOrderSearchDto(order.id,
-                        order.number,
-                        order.createDate,
-                        order.status,
-                        member.address.zipcode,
-                        member.address.street_address,
-                        member.address.jibeon_address,
-                        member.address.detail_address);
+        QOrderSearchDto qOrderSearchDto = new QOrderSearchDto(order.id, order.number, order.createDate, order.status, member.address.zipcode, member.address.street_address, member.address.jibeon_address, member.address.detail_address);
 
-        List<OrderSearchDto> content = queryFactory
-                .select(qOrderSearchDto)
-                .from(order)
-                .where(order.member.eq(findMember))
-                .orderBy(order.createDate.desc())
-                .join(order.member, member)
-                .fetch();
+        List<OrderSearchDto> content = queryFactory.select(qOrderSearchDto).from(order).where(order.member.eq(findMember)).orderBy(order.createDate.desc()).join(order.member, member).fetch();
 
         content.forEach(this::injectOrderDetails);
         return content;
@@ -60,17 +46,11 @@ public class OrderCustomImpl implements OrderCustom {
     }
 
     public JPAQuery<Long> totalCount(Member member) {
-        return queryFactory.select(order.count())
-                .from(order)
-                .where(order.member.eq(member));
+        return queryFactory.select(order.count()).from(order).where(order.member.eq(member));
     }
 
     private Order findById(Long orderId) {
-        return queryFactory
-                .select(order)
-                .from(order)
-                .where(order.id.eq(orderId))
-                .fetchOne();
+        return queryFactory.select(order).from(order).where(order.id.eq(orderId)).fetchOne();
     }
 
 }
