+++ src/test/java/shop/tryit/domain/answer/AnswerServiceTests.java
@@ -12,8 +12,8 @@ import org.springframework.data.domain.Page;
 import org.springframework.transaction.annotation.Transactional;
 import shop.tryit.domain.answer.entity.Answer;
 import shop.tryit.domain.answer.service.AnswerService;
-import shop.tryit.domain.member.Member;
-import shop.tryit.domain.member.MemberRepository;
+import shop.tryit.domain.member.entity.Member;
+import shop.tryit.domain.member.repository.MemberRepository;
 import shop.tryit.domain.question.QuestionRepository;
 import shop.tryit.domain.question.entity.Question;
 import shop.tryit.repository.answer.AnswerJpaRepository;
