+++ src/test/java/shop/tryit/domain/item/ItemRepositoryTests.java
@@ -11,8 +11,8 @@ import org.springframework.transaction.annotation.Transactional;
 import shop.tryit.domain.item.dto.ItemSearchDto;
 import shop.tryit.domain.item.entity.Item;
 import shop.tryit.domain.item.repository.ItemRepository;
-import shop.tryit.domain.member.Member;
-import shop.tryit.domain.member.MemberRepository;
+import shop.tryit.domain.member.entity.Member;
+import shop.tryit.domain.member.repository.MemberRepository;
 import shop.tryit.domain.order.entity.Order;
 import shop.tryit.domain.order.entity.OrderDetail;
 import shop.tryit.domain.order.entity.OrderStatus;
