+++ src/test/java/shop/tryit/domain/question/QuestionServiceTests.java
@@ -15,8 +15,8 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.transaction.annotation.Transactional;
 import shop.tryit.domain.answer.entity.Answer;
 import shop.tryit.domain.answer.repository.AnswerRepository;
-import shop.tryit.domain.member.Member;
-import shop.tryit.domain.member.MemberRepository;
+import shop.tryit.domain.member.entity.Member;
+import shop.tryit.domain.member.repository.MemberRepository;
 import shop.tryit.domain.question.dto.QuestionSearchCondition;
 import shop.tryit.domain.question.dto.QuestionSearchDto;
 import shop.tryit.domain.question.entity.Question;
