+++ src/test/java/shop/tryit/repository/question/QuestionRepositoryImplTests.java
@@ -12,8 +12,8 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.transaction.annotation.Transactional;
 import shop.tryit.domain.answer.entity.Answer;
 import shop.tryit.domain.answer.repository.AnswerRepository;
-import shop.tryit.domain.member.Member;
-import shop.tryit.domain.member.MemberRepository;
+import shop.tryit.domain.member.entity.Member;
+import shop.tryit.domain.member.repository.MemberRepository;
 import shop.tryit.domain.question.QuestionRepository;
 import shop.tryit.domain.question.dto.QuestionSearchCondition;
 import shop.tryit.domain.question.dto.QuestionSearchDto;
